+++ examples/nextjs-swc/src/app/[lang]/layout.tsx
@@ -1,0 +1,38 @@
+import linguiConfig from '../../../lingui.config'
+import { allMessages, getI18nInstance } from '../../appRouterI18n'
+import { LinguiClientProvider } from '../../components/LinguiClientProvider'
+import { PageLangParam, withLinguiLayout } from '../../withLingui'
+import React from 'react'
+import { t } from '@lingui/macro'
+
+export async function generateStaticParams() {
+  return linguiConfig.locales.map((lang) => ({ lang }))
+}
+
+export function generateMetadata({ params }: PageLangParam) {
+  const i18n = getI18nInstance(params.lang)
+
+  return {
+    title: t(i18n)`Translation Demo`
+  }
+}
+
+export default withLinguiLayout(function RootLayout({
+  children,
+  params: { lang }
+}) {
+  return (
+    <html lang={lang}>
+      <body className="bg-background text-foreground">
+        <main className="min-h-screen flex flex-col">
+          <LinguiClientProvider
+            initialLocale={lang}
+            initialMessages={allMessages[lang]!}
+          >
+            {children}
+          </LinguiClientProvider>
+        </main>
+      </body>
+    </html>
+  )
+})