+++ examples/remix-vite-babel/lingui.config.ts
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/export
+export * from "./app/modules/lingui/config";