+++ packages/cli/src/extract-experimental/bundleSource.ts
@@ -30,6 +30,7 @@ export async function bundleSource(
     "png",
     "css",
     "sass",
+    "scss",
     "less",
     "jpg",
   ]
@@ -49,7 +50,6 @@ export async function bundleSource(
     sourceRoot: outDir,
     sourcesContent: false,
     packages: "external",
-    outbase: rootDir,
     metafile: true,
 
     plugins: [