+++ packages/react/build.config.ts
@@ -1,0 +1,13 @@
+import { defineBuildConfig } from "unbuild"
+
+export default defineBuildConfig({
+  rollup: {
+    output: {
+      banner: (chunk: any) => {
+        if (chunk.name === "index") {
+          return `'use client';`
+        }
+      },
+    },
+  },
+})