+++ website/src/components/Header.tsx
@@ -42,14 +42,14 @@ const Header = (): React.ReactElement => {
               />
               <h1 className={clsx(styles.heroTitle)}>{siteConfig.tagline}</h1>
               <p className={"margin-bottom--lg"}>
-                JavaScript library for internalization (i18n) of JavaScript projects, including React, Vue, Node.js, and
-                Angular.
+                JavaScript library for internalization (i18n) of JavaScript projects. Supports React (including RSC and
+                React Native), Vue, Node.js, and Angular.
               </p>
 
               <div className={clsx(styles.heroButtons, "name", "margin-vert--md")}>
-                <Button href={useBaseUrl("/tutorials/setup-react")}>Get Started</Button>
-                <Button href={useBaseUrl("/introduction")} isOutline={true}>
-                  View Docs
+                <Button href={useBaseUrl("/introduction")}>Get Started</Button>
+                <Button href={useBaseUrl("/misc/examples")} isOutline={true}>
+                  Examples
                 </Button>
               </div>
               <iframe